requirements.txt
#aws-cdk-lib==2.44.0
aws-cdk-lib
# aws-cdk.aws-apigateway
# aws-cdk.aws-codedeploy
# aws-cdk.aws-codepipeline
# aws-cdk.aws-codepipeline-actions
# aws-cdk.aws-lambda
# aws-cdk.core
# aws-cdk.pipelines
#constructs>=10.0.0,<11.0.0 
constructs
